[REF] l10n_br_fiscal: merge document.mixin.methods into document.mixin
